Migrate to jest

+++ src/test/e2e/api/admin/addon.e2e.test.js
@@ -1,7 +1,5 @@
 'use strict';
 
-const test = require('ava');
-
 const dbInit = require('../../helpers/database-init');
 const { setupApp } = require('../../helpers/test-helper');
 const getLogger = require('../../../fixtures/no-logger');
@@ -11,31 +9,33 @@ const MASKED_VALUE = '*****';
 let stores;
 let db;
 
-test.before(async () => {
+beforeAll(async () => {
     db = await dbInit('addon_api_serial', getLogger);
     stores = db.stores;
 });
 
-test.after.always(async () => {
-    await db.destroy();
+afterAll(async () => {
+    if (db) {
+        await db.destroy();
+    }
 });
 
-test.serial('gets all addons', async t => {
-    t.plan(3);
+test('gets all addons', async () => {
+    expect.assertions(3);
     const request = await setupApp(stores);
     return request
         .get('/api/admin/addons')
         .expect('Content-Type', /json/)
         .expect(200)
         .expect(res => {
-            t.is(res.body.addons.length, 0, 'expected 0 configured addons');
-            t.is(res.body.providers.length, 4, 'expected 4 addon providers');
-            t.is(res.body.providers[0].name, 'webhook');
+            expect(res.body.addons.length).toBe(0);
+            expect(res.body.providers.length).toBe(4);
+            expect(res.body.providers[0].name).toBe('webhook');
         });
 });
 
-test.serial('should not be able to create invalid addon', async t => {
-    t.plan(0);
+test('should not be able to create invalid addon', async () => {
+    expect.assertions(0);
     const request = await setupApp(stores);
     return request
         .post('/api/admin/addons')
@@ -43,8 +43,8 @@ test.serial('should not be able to create invalid addon', async t => {
         .expect(400);
 });
 
-test.serial('should create addon configuration', async t => {
-    t.plan(0);
+test('should create addon configuration', async () => {
+    expect.assertions(0);
     const request = await setupApp(stores);
 
     const config = {
@@ -63,8 +63,8 @@ test.serial('should create addon configuration', async t => {
         .expect(201);
 });
 
-test.serial('should delete addon configuration', async t => {
-    t.plan(0);
+test('should delete addon configuration', async () => {
+    expect.assertions(0);
     const request = await setupApp(stores);
 
     const config = {
@@ -86,8 +86,8 @@ test.serial('should delete addon configuration', async t => {
     await request.delete(`/api/admin/addons/${id}`).expect(200);
 });
 
-test.serial('should update addon configuration', async t => {
-    t.plan(2);
+test('should update addon configuration', async () => {
+    expect.assertions(2);
     const request = await setupApp(stores);
 
     const config = {
@@ -125,16 +125,15 @@ test.serial('should update addon configuration', async t => {
         .send(config)
         .expect(200)
         .expect(r => {
-            t.is(r.body.parameters.url, MASKED_VALUE);
-            t.is(
-                r.body.parameters.bodyTemplate,
+            expect(r.body.parameters.url).toBe(MASKED_VALUE);
+            expect(r.body.parameters.bodyTemplate).toBe(
                 updatedConfig.parameters.bodyTemplate,
             );
         });
 });
 
-test.serial('should not update with invalid addon configuration', async t => {
-    t.plan(0);
+test('should not update with invalid addon configuration', async () => {
+    expect.assertions(0);
     const request = await setupApp(stores);
 
     const config = {
@@ -152,8 +151,8 @@ test.serial('should not update with invalid addon configuration', async t => {
         .expect(400);
 });
 
-test.serial('should not update unknown addon configuration', async t => {
-    t.plan(0);
+test('should not update unknown addon configuration', async () => {
+    expect.assertions(0);
     const request = await setupApp(stores);
 
     const config = {
@@ -172,8 +171,8 @@ test.serial('should not update unknown addon configuration', async t => {
         .expect(404);
 });
 
-test.serial('should get addon configuration', async t => {
-    t.plan(3);
+test('should get addon configuration', async () => {
+    expect.assertions(3);
     const request = await setupApp(stores);
 
     const config = {
@@ -197,24 +196,23 @@ test.serial('should get addon configuration', async t => {
         .get(`/api/admin/addons/${id}`)
         .expect(200)
         .expect(r => {
-            t.is(r.body.provider, config.provider);
-            t.is(
-                r.body.parameters.bodyTemplate,
+            expect(r.body.provider).toBe(config.provider);
+            expect(r.body.parameters.bodyTemplate).toBe(
                 config.parameters.bodyTemplate,
             );
-            t.is(r.body.parameters.url, MASKED_VALUE);
+            expect(r.body.parameters.url).toBe(MASKED_VALUE);
         });
 });
 
-test.serial('should not get unknown addon configuration', async t => {
-    t.plan(0);
+test('should not get unknown addon configuration', async () => {
+    expect.assertions(0);
     const request = await setupApp(stores);
 
     await request.get('/api/admin/addons/445').expect(404);
 });
 
-test.serial('should not delete unknown addon configuration', async t => {
-    t.plan(0);
+test('should not delete unknown addon configuration', async () => {
+    expect.assertions(0);
     const request = await setupApp(stores);
 
     return request.delete('/api/admin/addons/21231').expect(404);
