+++ src/test/e2e/api/admin/event.e2e.test.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const test = require('ava');
 const { setupApp } = require('../../helpers/test-helper');
 const dbInit = require('../../helpers/database-init');
 const getLogger = require('../../../fixtures/no-logger');
@@ -8,17 +7,19 @@ const getLogger = require('../../../fixtures/no-logger');
 let stores;
 let db;
 
-test.before(async () => {
+beforeAll(async () => {
     db = await dbInit('event_api_serial', getLogger);
     stores = db.stores;
 });
 
-test.after.always(async () => {
-    await db.destroy();
+afterAll(async () => {
+    if (db) {
+        await db.destroy();
+    }
 });
 
-test.serial('returns events', async t => {
-    t.plan(0);
+test('returns events', async () => {
+    expect.assertions(0);
     const request = await setupApp(stores);
     return request
         .get('/api/admin/events')
@@ -26,8 +27,8 @@ test.serial('returns events', async t => {
         .expect(200);
 });
 
-test.serial('returns events given a name', async t => {
-    t.plan(0);
+test('returns events given a name', async () => {
+    expect.assertions(0);
     const request = await setupApp(stores);
     return request
         .get('/api/admin/events/myname')
